refactor: new tests suite structure
This commit changes drastically the testing suite structure we have.
Specifically, it made the next changes:

- Now the tests directory is a 1-1 mirror of the directories at cryptographic_estimators. Each estimator will have his main test handler file (named `test_<estimator_name>.py`), and could have a directory with legacy_implementations that we want to reuse as tests.

- Split a file that tested multiple estimators simultaneously on their respective estimators: `tests/test_estimators.py`
- Change the undescriptive directory name `module` to `helpers` and update all import references to its files.

- Clean the directory structure of the MQEstimator (legacy implementation).

- Improve import declarations and general code style via Black.
