+++ tests/MQEstimator/legacy_implementations/mpkc/series/nmonomial.py
