+++ tests/helpers/attack_cost.sage
