+++ tests/helpers/kmp_cost.sage
