+++ tests/helpers/our_cost.sage
